% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{as.data.frame.ts}
\alias{as.data.frame.ts}
\title{Transform time series object to data.frame}
\usage{
\method{as.data.frame}{ts}(x)
}
\arguments{
\item{x}{The time series of class "ts".}
}
\value{
A data.frame with the data from x in columns 2 onwards, and the
first column for the dates.
}
\description{
Represents an object of class "ts" as a data.frame. It accepts monthly,
quarterly, and annual series.
}
\details{
The first column of the resulting data.frame is a column of dates. For example,
for quarterly series, quarters are represented by the first month of the
quarter, e.g. 2015 Q2 turns into 01/04/2015.
}
\examples{
set.seed(23)
x <- sample(98:102, 10, TRUE) |> ts(start = 2010, frequency = 4)
as.data.frame(x)
}
