% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{df_to_ts}
\alias{df_to_ts}
\title{dataframe to time series}
\usage{
df_to_ts(df)
}
\arguments{
\item{df}{A dataframe whose first column are dates.}
}
\value{
A ts object.
}
\description{
Converts a dataframe to a ts object. Only time frequencies between
1 day and 1 year are supported.
}
\examples{
passengers <- datasets::AirPassengers
passengers_df <- ts_to_df(passengers)
df_to_ts(passengers_df)
}
