% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesTools-package.R
\docType{package}
\name{TimeSeriesTools-package}
\alias{TimeSeriesTools}
\alias{TimeSeriesTools-package}
\title{TimeSeriesTools: Helping Functions for Working with Time Series}
\description{
Helping functions for working with time series, such as a `as.data.frame()` function that adds a date column, and a function to compute the growth rate series at different lags.
}
\keyword{internal}
