% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ts_from_excel}
\alias{ts_from_excel}
\title{Read time series from an Excel file}
\usage{
ts_from_excel(filepath, ...)
}
\arguments{
\item{filepath}{(character) A path to an Excel file.}

\item{...}{Parameters passed to \code{read_excel()}}
}
\value{
A ts object.
}
\description{
Reads time series from an Excel file. It is assumed that the first column
are dates and the rest are numeric.
}
\details{
This function uses \code{readxl::read_excel()} to read the file, specifying that the
first column be of type date and the rest numeric. It reads the file twice
behind the scenes, to count the number of columns.

After the contents are loaded into a dataframe, \code{df_to_ts()} is applied.
}
\examples{
passengers_df <- ts_to_df(datasets::AirPassengers)
df_to_ts(passengers_df)
}
