% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ts_to_df}
\alias{ts_to_df}
\title{time series to dataframe}
\usage{
ts_to_df(x, date_name = "Date")
}
\arguments{
\item{x}{A time series object ts}

\item{date_name}{Name of the date column. Defaults to "Date".}
}
\value{
A dataframe with same column names (if mts) plus a first
column of dates
}
\description{
Converts a ts object to a dataframe whose first column are the dates.
}
\examples{
passengers <- datasets::AirPassengers
ts_to_df(passengers)
}
