% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_ts}
\alias{update_ts}
\title{Update time series}
\usage{
update_ts(x, y, update_shared = FALSE)
}
\arguments{
\item{x}{(ts/mts) Original time series.}

\item{y}{(ts/mts) New time series}

\item{update_shared}{(bool) If x and y don't have the same columns,
should the matching ones be updated? Default: FALSE.}
}
\description{
Takes a ts (mts) object and updates its data in place with a second ts (mts) object.
}
