% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_ts_uni}
\alias{update_ts_uni}
\title{Update (univariate) time series}
\usage{
update_ts_uni(x, y)
}
\arguments{
\item{x}{(ts/mts) Original time series.}

\item{y}{(ts/mts) New time series}
}
\description{
Takes a ts object and updates its data in place with a second ts object.
}
