% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{compute_gr}
\alias{compute_gr}
\title{Compute the growth-rate series}
\usage{
compute_gr(x, s)
}
\arguments{
\item{x}{A time series of class "ts"}

\item{s}{Lag at which the growth-rate is computed}
}
\value{
Series of growth-rates
}
\description{
Function that computes the growth-rate series of a given time series.
}
\examples{
set.seed(23)
x <- sample(98:102, 10, TRUE) |> ts(start = 2010, frequency = 4)
compute_gr(x, 4)
compute_gr(x, 1)
}
